Add files via upload
This commit includes the raw data obtained from baseball-reference.com 
The datasets include batting and pitching stats from 2015-2019. Batting and Pitching files can be identified by the ending of each file, where those ending in _B are for batting stats and thos ending in _P are for pitching stats.
The file "MLB_w_dummy" is a file created personally with data from the MLB official webpage and it includes a dummy for whether a team belongs to the American League (dummy == 1) or to the National League (dummy == 0).